Move rrdp content warnings to different file
